change badge color naming and refactor components
