+++ client/src/components/preferences/preferences-inputs/preferences-item-badge.tsx
@@ -1,6 +1,7 @@
 import { Badge } from "@/components/ui/badge";
 import { useRemovePreferenceItem } from "../mutation/use-remove-preferences.mutation";
 import { PlusIcon } from "lucide-react";
+import { useState } from "react";
 
 interface PreferenceItemProps {
   item: string;
@@ -9,23 +10,28 @@ interface PreferenceItemProps {
   removeItemMutation: ReturnType<typeof useRemovePreferenceItem>;
 }
 
-export function PreferenceItem({
+export function PreferenceItemBadge({
   item,
   index,
   handleRemoveItem,
   removeItemMutation,
 }: PreferenceItemProps) {
+  const [isButtonHovered, setIsButtonHovered] = useState(false);
   return (
     <Badge
       key={index}
       variant="secondary"
-      className="bg-secondary text-sidebar-foreground flex items-center gap-2"
+      className={`bg-primary text-white flex items-center gap-2 ${
+        isButtonHovered ? "bg-primary/80 " : "bg-primary"
+      } transition-colors duration-200`}
     >
       {item}
       <button
-        className="cursor-pointer text-sidebar-foreground hover:text-destructive"
+        className="cursor-pointer text-white hover:scale-125 transition-transform duration-200"
         onClick={() => handleRemoveItem(item)}
         disabled={removeItemMutation.isPending}
+        onMouseEnter={() => setIsButtonHovered(true)}
+        onMouseLeave={() => setIsButtonHovered(false)}
       >
         <PlusIcon className="h-3 w-3 rotate-45" />
       </button>
