+++ client/src/components/preferences/preferences-inputs/preferences-list.tsx
@@ -1,22 +1,34 @@
+import { usePreferences } from "@/hooks/query/use-preferences.query";
 import { useRemovePreferenceItem } from "../mutation/use-remove-preferences.mutation";
-import { PreferenceItem } from "./preferences-item";
+import { PreferenceArrayKeys } from "../preferences.type";
+import { PreferenceItemBadge } from "./preferences-item-badge";
 
 interface PreferenceListProps {
-  items: string[];
-  handleRemoveItem: (item: string) => void;
-  removeItemMutation: ReturnType<typeof useRemovePreferenceItem>;
+  preferenceKey: keyof PreferenceArrayKeys;
 }
 
-export function PreferenceList({
-  items,
-  handleRemoveItem,
-  removeItemMutation,
-}: PreferenceListProps) {
+export function PreferenceList({ preferenceKey }: PreferenceListProps) {
+  const removeItemMutation = useRemovePreferenceItem();
+  const { data, isPending } = usePreferences();
+
+  if (isPending) {
+    return <div className="text-gray-500">Loading...</div>;
+  }
+
+  const items = data?.[preferenceKey] || [];
+
+  const handleRemoveItem = async (item: string) => {
+    await removeItemMutation.mutateAsync({
+      field: preferenceKey,
+      item,
+    });
+  };
+
   return (
     <div className="flex  flex-wrap gap-2">
       {items.length > 0 ? (
         items.map((item, index) => (
-          <PreferenceItem
+          <PreferenceItemBadge
             key={index}
             item={item}
             index={index}
